Refactor MixedOp classes
